Constraints classes, input modes, restructuring
Created general class for constraint types & restructured some of the
calculateConstraints method to leverage this.  Also restructured the
MyWorld class to add the support for them

Also created support for different modes of input by the user, and added
a menu to display the controls to activate them.  These include:
- Select : user selects a particle
- Insert : user inserts a particle
- Spring : user applies a spring force to selected particle